+++ src/pages/Assets/index.js
@@ -1,12 +1,12 @@
 import React from 'react'
 import { Redirect } from 'react-router-dom'
-import { Section } from '../Watchlists/New'
+import Section from '../Watchlists/Section'
 import Page from '../../ducks/Page'
 import StoriesList from '../../components/Stories/StoriesList'
 import Trends from '../../components/Trends/Trends'
 import RecentlyWatched from '../../components/RecentlyWatched/RecentlyWatched'
 import FeaturedWatchlistCards from '../../ducks/Watchlists/Featured'
-import styles from '../Watchlists/New.module.scss'
+import styles from '../Watchlists/index.module.scss'
 
 const Assets = ({ isDesktop }) => {
   if (isDesktop) {
@@ -17,7 +17,7 @@ const Assets = ({ isDesktop }) => {
     <Page title='Explore assets' className={styles.wrapper}>
       <StoriesList classes={styles} />
 
-      <RecentlyWatched className={styles.recents} type='assets' />
+      <RecentlyWatched type='assets' />
 
       <Section title='Indices'>
         <FeaturedWatchlistCards className={styles.card} />
